% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_earliest.R
\name{dist2arc}
\alias{dist2arc}
\title{Approximate angle for a distance}
\usage{
dist2arc(x, dm)
}
\arguments{
\item{x}{A \code{sfc_POINT} object in longlat coordinates,
indicating a position in the globe.}

\item{dm}{A target distance in m.}
}
\value{
A numeric value of angle in decimal degrees, such that a
  displacement in longitude and latitude by that magnitude
  corresponds to approximately the given distance.
}
\description{
Compute an approximate angle that correspond to a given distance
in both longitude and latitude.
}
\details{
The distance corresponding to an angle in longitude and latitude
depends on the location in the earth and is typically different
for both directions.

This function computes the angle that yields approximately the
target distance when added to the longitude or the latitude by
minimising the squared errors.

The average error is returned as an attribute.
}
\examples{

## What angle yields 1 km at the equator? (and with which precision?)
x0 <- st_sfc(st_point(c(0, 0)), crs = 4326)
dist2arc(x0, 1e3)

## Note that this is independent of the longitude
x1 <- st_sfc(st_point(c(90, 0)), crs = 4326)
dist2arc(x1, 1e3)

## But is strongly dependent of the latitude
x2 <- st_sfc(st_point(c(0, 45)), crs = 4326)
x3 <- st_sfc(st_point(c(0, 75)), crs = 4326)
dist2arc(x2, 1e3)
dist2arc(x3, 1e3)
}
