% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mask.R
\name{estimation_mask}
\alias{estimation_mask}
\title{Raster termplate for estimation}
\usage{
estimation_mask(x, buffer_size, res = buffer_size/8)
}
\arguments{
\item{x}{An object of class \code{sfg}, \code{sfc} or \code{sf},
including particularly objects of class \code{sr_obs}. Geographical
objects that determine the estimation region.}

\item{buffer_size}{Numeric. Expansion of the extent around the
convex-hull of \code{x}. In the units of the data (e.g. m, or
arc-degrees).}

\item{res}{Numeric vector of length 1 or 2. Resolution of the raster
template. Size(s) of the pixel, in metres.}
}
\description{
Given a set of observations, define a estimation template.
}
\examples{

## Some random points in the unit square
pp <- st_multipoint(matrix(runif(30), ncol = 2))
(em <- estimation_mask(pp, buffer_size = .3, res = .1))
plot(em)
points(st_coordinates(pp))
}
