% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_earliest.R
\name{filter_earliest_neigh}
\alias{filter_earliest_neigh}
\title{Earliest observations within a neighbourhood}
\usage{
filter_earliest_neigh(x)
}
\arguments{
\item{x}{sr_obs object.}

\item{dist}{Distance parameter in m.}
}
\value{
Another sr_obs object with a subset of the points and
the earliest times observed in the neighbourhood of each.
}
\description{
Select a subset of the observation points at least \code{tol} m far
apart as representative of the emph{neighbourhood}. Assign the
earliest times observed in the neighbourhood to them.
}
\details{
Note that the neighbouring distance must be expressed in m, even
when the coordinates are geographical.

This function reproduces the operation for all the Monte Carlo
replicates if any.

For the original dataset, the neighbouring tolerance parameter
is the mean value of the interval specified in its \code{uq}.
}
\examples{
  d <- data.frame(lon = runif(30), lat = runif(30), date = 1:30)

  ## Use between 10 and 20 \% of data diameter as neighbouring
  ## tolerance. The main result will use exactly 15\% while the
  ## \\code{mc} replicates will use random Gaussian values from
  ## that interval (at 99.9\%)
  sruq <- sr_uq(10, 0, 1, neigh_tol = c(-10, -20))
  sro <- sr_obs(d, "date", uq = sruq)
  srf <- filter_earliest_neigh(sro)
}
