% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_fdoi.R
\name{fit_surface}
\alias{fit_surface}
\title{Fit a Thin-plate splines model to a set of observations}
\usage{
fit_surface(x)
}
\arguments{
\item{x}{A \code{sr_obs} object defined with \code{\link{sr_obs}}.}
}
\value{
An object of class \code{Krig} and \code{Tps} which includes
the fitted values and the model residuals. See
\code{\link[fields]{Tps}}. It also allows interpolation using
\code{\link[raster]{interpolate}}.
}
\description{
Fit a Thin-plate splines model to a set of observations
}
\examples{
d <- data.frame(lon = runif(30), lat = runif(30))
d <- transform(d, date = round(10 * sqrt(lon**2 + lat**2)))
sro <- sr_obs(d, "date")
fm <- fit_surface(sro)
summary(fm)
plot(fm)    # Some model diagnostics plots
fields::surface(fm)  # Quick image/contour plot of predicted surface.
}
