% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lonlatvars.R
\name{get_lonlatvars}
\alias{get_lonlatvars}
\title{Identify variables with geographical coordinates}
\usage{
get_lonlatvars(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
A character vector with the variable names for longitude
  and latitude, or an error.
}
\description{
Given a data.frame with two variables with longitudes and latitudes,
return the variable names.
}
\details{
This function explores the names of the variables for the patterns
'lon' or 'long' and 'lat' ignoring case. If exactly one of each
patterns is detected, it verifies that their magnitudes are within
-365 and 365 for the longitude and within -90.1 and 90.1 for the
latitude.
}
\examples{
  d <- data.frame(lon = 1:10, lat = 1:10)
  get_lonlatvars(d)
}
