% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invslope.R
\name{invslope}
\alias{invslope}
\title{Compute the inverse slope of the input surface}
\usage{
invslope(x, rmtop, bnd)
}
\arguments{
\item{x}{\code{Raster*} with estimated first-arrival date. Needs to be
projected.}

\item{rmtop}{Numeric, between 0 and 100. Discard the \code{rmtop} top percent
of values.}

\item{bnd}{Numeric vector of length 2. Expected range of values. Everything
beyond this range is filtered.}
}
\value{
An object of the same class \code{Raster*} as the input.
}
\description{
This function is used to derive local spread rate from a
first-date-of-invasion surface
}
\details{
To control for spurious numerical divergences, the user can remove
the largest \code{rmtop} % of the values, and all those
values beyond a reasonable interval \code{bnd}.
}
\examples{

## A flat surface with constant slope of 1
r_values <- replicate(100, seq(0.01, 1, length = 100))
r <- raster::raster(r_values, crs = sp::CRS("+init=epsg:3857"))
raster::plot(r)
invslope(r)
}
