% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_obs.R
\name{mc_sample}
\alias{mc_sample}
\title{Monte Carlo samples from a dataset}
\usage{
mc_sample(x)
}
\arguments{
\item{x}{An object of class \code{sr_obs} from function
\code{\link{sr_obs}}}
}
\description{
Shuffle observations in space and time.
}
\details{
Add an attribute \code{mc} with a list of datasets with some noise
added to the original coordinates and dates. The number of Monte
Carlo samples and the variablity of the noise is given by the
\code{sr_uq} attribute in the object.
}
\examples{
  d <- data.frame(lon = 1, lat = 1, date = 1)
  sr_obs(d, "date", uq = sr_uq(nsim = 3, space = 1, time = 1))
}
