% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_earliest.R
\name{representative_points}
\alias{representative_points}
\title{Choose representative points}
\usage{
representative_points(x, dTolerance)
}
\arguments{
\item{x}{Object of class \code{sf}}

\item{dTolerance}{Numeric. Distance tolerance in m.}
}
\value{
Object of class \code{sf} with a subset of the original
locations in \code{x}.
}
\description{
Select a subset of points at a minimum tolerance distance from
each other.
}
\examples{

## 50 points in the unit square
require(sf)
x <- st_sfc(st_multipoint(matrix(runif(100), ncol = 2)), crs = 3857)
representative_points(x, .5)
}
