% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{sr}
\alias{sr}
\title{Spread rate}
\usage{
sr(x, r = estimation_mask(x, buffer_size = max(st_distance(x))/10))
}
\arguments{
\item{x}{A \code{sr_obs} object defined with \code{\link{sr_obs}}.}

\item{r}{A \code{raster} template for results.}
}
\description{
Estimate the local spread rate of an epidemiological invasion.
}
\details{
This function will compute spread-rate estimates for all of the
Monte Carlo samples in the dataset. It uses internally the function
\code{\link[furrr]{future_map}} which will take advantage of
multiple processors or cluster access if you set a proper
\code{\link[future]{plan}} beforehand. See examples.
}
\examples{

require(sf)
## Randomly sample 50 point in the unit square
x_mat <- matrix(runif(100), ncol = 2)

## Make it spatial point in a projected CRS
x_sfc <- st_cast(st_sfc(st_multipoint(x_mat), crs = 3857), "POINT")

## Assign dates proportional to the squared distance to the origin
## So that spread-rate decreases linearly
sro <- sr_obs(st_sf(x_sfc, date = sqrt(rowSums(x_mat**2))), "date")
sro <- sr_obs(st_sf(x_sfc, date = rowSums(x_mat**2)), "date",
uq = sr_uq(6, .2, .2, .3))

## Estimate local spread-rate
## in the units of the coordinates divided by the units of time
sre <- sr(sro)
raster::plot(sre, col = hcl.colors(12))
points(st_coordinates(sro))

}
