% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_fdoi.R
\name{sr_fdoi}
\alias{sr_fdoi}
\title{First-date of invasion}
\usage{
sr_fdoi(x, estimation_mask)
}
\arguments{
\item{x}{A \code{sr_obs} object defined with \code{\link{sr_obs}}.}

\item{estimation_mask}{A \code{raster} template for results.}
}
\value{
A RasterLayer with estimated dates of invasion at each
  pixel.
}
\description{
Estimate a surface of first-date of invasion, given a sparse set
of observation locations and times.
}
\details{
This function first takes the earliest observed case in a neighbourhood,
as defined by the parameter \code{neigh_tol} in the definition of
\code{x}. Next, it interpolates the earliest observations with a
smooth surface, using a Thin-Plate splines model.
}
\examples{
d <- data.frame(lon = runif(30), lat = runif(30))
d <- transform(d, date = round(10 * sqrt(lon**2 + lat**2)))
sro <- sr_obs(d, "date")
r <- raster(st_sf(sro), vals = 1)
fdoi <- sr_fdoi(sro, r)
plot(fdoi, col = hcl.colors(12))
points(d)
}
