% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_obs.R
\name{sr_obs}
\alias{sr_obs}
\title{Define a observations dataset}
\usage{
sr_obs(x, timevar, uq)
}
\arguments{
\item{x}{\code{data frame} with geographical coordinates or
\code{sf} object of type POINT.}

\item{timevar}{Character. Variable name with observation times or
dates.}

\item{uq}{Object created with \code{sr_uq()} with parameters of
Uncertainty Quantification.}
}
\value{
Object of class \code{sr_obs}, which is a \code{sf} of type
  POINT with complementary meta-data.
}
\description{
Provide a dataset with the necessary meta-data for spread-rate
estimation.
}
\details{
If your dataset is projected (no lon/lat variables) then first
build the \code{sf} object yourself using the corresponding
Coordinate Reference System. See example.

This function will produce the Monte Carlo samples from the
dataset and the \code{uq} object, if necessary. It uses internally
the function \code{\link[furrr]{future_map}} which will take advantage
of multiple processors or cluster access if you set a proper
\code{\link[future]{plan}} beforehand. See examples.
}
\examples{
  d <- data.frame(lon = 1, lat = 1, date = 1)
  sr_obs(d, "date")

  ## Projected coordinates
  d <- data.frame(x = 1, y = 1, date = 1)
  d.sf <- sf::st_as_sf(d, coords = c("x", "y"), crs = 27561)
  sr_obs(d.sf, "date")

\dontrun{
  ## Perform Monte Carlo samples in parallel
  library(furrr)
  plan(multiprocess)
  uq <- sr_uq(nsim = 3, space = 1, time = 1)
  sr_obs(d, "date", uq = uq)
}

}
