% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{sr_uq}
\alias{sr_uq}
\title{Parameters for Uncertainty Quantification}
\usage{
sr_uq(nsim = 1L, space = 0, time = 0, neigh_tol = -2.5)
}
\arguments{
\item{nsim}{Integer > 0. Number of Monte Carlo replicates.}

\item{space}{Numeric, non-negative. Uncertainty of spatial coordinates.}

\item{time}{Numeric, non-negative. Uncertainty of temporal values.}

\item{neigh_tol}{Number or interval (numeric vector of length 2).
Neighouring-tolerance parameter in the units of the dataset
coordinates if positive, or relative to the dataset diameter if
negative.}
}
\value{
An object of class \code{sr_uq} which is a list with the
  given or default values, after some sanity checks.
}
\description{
Define the parameters for quantifying the uncertainty in the
estimation of the spread-rate.
}
\details{
The actual observations will be The number of Monte Carlo replicates
The \code{space} and \code{time} arguments are expressed in the
data spatial and temporal units.

For each one of the \code{nsim} replicates, the observations will
be randomly shifted in space within a circle of radious
\code{space}, while the dates will be shifted within the interval
+- \code{time} with a discrete approximation to a Gaussian
curve.

The {neighbouring-tolerance} parameter \code{neigh_tol} is used to
filter the earliest observed cases in a \emph{neighbourhood}. If
two points are closer than this value, they are considered roughly
in the same neighbourhood. The date of invasion of a location is
the earliest observed case in the neighbourhood. It can be defined
as a numeric value in the units of the coordinates, or as a range
(i.e. a numeric vector of length 2). Negative values will be
interpreted as a percentage of the diameter of the dataset.
}
\examples{
  ## Use the default values: work with current observations
  ## without quantifying uncertainty
  uq_def <- sr_uq()

  ## Set the neighbouring-tolerance parameter only
  uq1 <- sr_uq(neigh_tol = 800)

  ## 10 MC replicates, shift locations within a circle of radious
  ## 1 km, shift dates by zero/one day up or down, consider points
  ## closer than 5\% of the diameter of the dataset as in the same
  ## neighbourhood
  uq <- sr_uq(nsim = 10, space = 1e3, time = 1, neigh_tol = -5)
}
