% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_transform.R
\name{st_transform.sr_obs}
\alias{st_transform.sr_obs}
\title{Transform coordinates of a sr_obs object}
\usage{
\method{st_transform}{sr_obs}(x, crs, ...)
}
\arguments{
\item{x}{A \code{sr_obs} object created with \code{\link{sr_obs}}.}

\item{crs}{Coordinate reference system: inteer with the EPSG code
or character with proj4string}

\item{...}{Ignored}
}
\value{
A \code{sr_obs} object in the target projection.
}
\description{
Transform the underlying \code{sf} object and all the MC replicas,
if any.
}
\examples{
## Some observed data and MC replicates in longitude/latitude
x <- sr_obs(data.frame(lat = 1, lon = 1, day = 1), "day", uq = sr_uq(2, 1, 1))
st_crs(x)
lapply(attr(x, "mc"), st_crs)

xt <- st_transform(x, 3857)
st_crs(xt)
lapply(attr(xt, "mc"), st_crs)
}
