% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_jitter.R
\name{time_jitter}
\alias{time_jitter}
\title{Jitter a vector of 1D values}
\usage{
time_jitter(x, e)
}
\arguments{
\item{x}{Numeric or Date vector.}

\item{e}{Numeric. Semi-width of the jittering interval.}
}
\value{
Vector of the same size and type as \code{x} with jittered
values.
}
\description{
Modify a vector of dates by a random quantity in [-e, e].
The distribution decreases exponentially, with rate 1/2
}
\examples{
spreadrate:::time_jitter(1:10, 1)  # Numeric input
spreadrate:::time_jitter(Sys.Date() + 1:10, 1)  # Date input

## Distribution of jittering
x <- seq.int(1e3)
xj <- spreadrate:::time_jitter(x, 5)
barplot(table(x - xj))
}
